--- src/app/admin/login/actions.ts
@@ -1,58 +1,0 @@
-// src/app/admin/login/actions.ts
-'use server'
-
-import { revalidatePath } from 'next/cache'
-import { redirect } from 'next/navigation'
-import { createClient } from '@/utils/supabase/server'
-
-export async function login(formData: FormData) {
-  const supabase = await createClient()
-
-  const email = formData.get('email') as string
-  const password = formData.get('password') as string
-
-  console.log('Login attempt:', { email, hasPassword: !!password })
-
-  // Validate inputs
-  if (!email || !password) {
-    console.error('Missing email or password')
-    redirect('/admin/login?error=Please provide both email and password')
-  }
-
-  const { data, error } = await supabase.auth.signInWithPassword({
-    email,
-    password,
-  })
-
-  if (error) {
-    console.error('Login error:', error.message)
-    redirect(`/admin/login?error=${encodeURIComponent(error.message)}`)
-  }
-
-  console.log('Login successful for:', email)
-  revalidatePath('/', 'layout')
-  redirect('/admin/dashboard')
-}
-
-export async function signup(formData: FormData) {
-  const supabase = await createClient()
-
-  const email = formData.get('email') as string
-  const password = formData.get('password') as string
-
-  if (!email || !password) {
-    redirect('/admin/login?error=Please provide both email and password')
-  }
-
-  const { error } = await supabase.auth.signUp({
-    email,
-    password,
-  })
-
-  if (error) {
-    redirect(`/admin/login?error=${encodeURIComponent(error.message)}`)
-  }
-
-  revalidatePath('/', 'layout')
-  redirect('/admin/dashboard')
-}
